requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
charset-normalizer==2.1.1
discord==2.0.0
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
pycodestyle==2.9.1
python-dotenv==0.21.0
tomli==2.0.1
yarl==1.8.1
